requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
mysql-connector-python==8.3.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pandas==2.2.0
reportlab==4.0.9
python-dotenv==1.0.0
pydantic==2.5.3